+++ packages/mint-components/src/components/sqm-widget-verification/sqm-widget-verification.tsx
@@ -1,77 +1,18 @@
 import {
   isDemo,
-  useLazyQuery,
   useParentState,
   useSetParent,
-  useUserIdentity,
 } from "@saasquatch/component-boilerplate";
-import { useState, withHooks } from "@saasquatch/stencil-hooks";
+import { withHooks } from "@saasquatch/stencil-hooks";
 import { Component, h, Prop } from "@stencil/core";
+import { getProps } from "../../utils/utils";
+import { extractProps } from "../tax-and-cash/sqm-tax-and-cash/extractProps";
 import {
   SHOW_CODE_NAMESPACE,
   VERIFICATION_EMAIL_NAMESPACE,
   VERIFICATION_PARENT_NAMESPACE,
 } from "./keys";
-import { getProps } from "../../utils/utils";
-import { extractProps } from "../tax-and-cash/sqm-tax-and-cash/extractProps";
-import { gql } from "graphql-request";
-import { useEffect } from "@saasquatch/universal-hooks";
-
-const USER_LOOKUP = gql`
-  query checkUserVerification {
-    viewer {
-      ... on User {
-        id
-        accountId
-        email
-        emailVerified
-        managedIdentity {
-          email
-          emailVerified
-        }
-      }
-    }
-  }
-`;
-
-function useWidgetVerificationInternal() {
-  const userIdentity = useUserIdentity();
-  const [showCode, setShowCode] = useParentState<boolean>({
-    namespace: SHOW_CODE_NAMESPACE,
-    initialValue: false,
-  });
-  const [email, setEmail] = useParentState<string | undefined>({
-    namespace: VERIFICATION_EMAIL_NAMESPACE,
-    initialValue: userIdentity?.email,
-  });
-  const setContext = useSetParent(VERIFICATION_PARENT_NAMESPACE);
-  const [loading, setLoading] = useState(true);
-  const [fetch] = useLazyQuery(USER_LOOKUP);
-
-  useEffect(() => {
-    const checkUser = async () => {
-      try {
-        const res = await fetch({});
-        if (!res || res instanceof Error) throw new Error();
-
-        if (res?.viewer?.emailVerified) setContext(true);
-        else if (res?.viewer?.managedIdentity?.emailVerified) setContext(true);
-      } catch (e) {
-        console.error("Could not fetch user information:", e);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    checkUser();
-  }, []);
-
-  const onVerification = () => {
-    setContext(true);
-  };
-
-  return { showCode, onVerification, loading };
-}
+import { useWidgetVerification } from "./useWidgetVerification";
 
 /**
  * @uiName Widget Verification Flow
@@ -211,7 +152,7 @@ export class WidgetVerification {
   render() {
     const { showCode, onVerification, loading } = isDemo()
       ? useDemoWidgetVerificationInternal()
-      : useWidgetVerificationInternal();
+      : useWidgetVerification();
 
     if (loading) return <sl-spinner></sl-spinner>;
