+++ packages/mint-components/src/index.html
@@ -1,30 +1,16 @@
 <!DOCTYPE html>
 <html dir="ltr" lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0"
-    />
-    <title>Mint Components Stencilbook</title>
-    <script type="module" src="build/mint-components.esm.js"></script>
-    <script nomodule src="build/mint-components.js"></script>
-    <script src="http://localhost:3000/dist/bedrock-components/bedrock-components.esm.js"></script>
-  </head>
 
-  <body>
-    <sqm-stencilbook></sqm-stencilbook>
-    <!-- <sqm-widget-verification>
-      <template slot="verified"><h1>VERIFIED</h1></template>
-    </sqm-widget-verification>
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0" />
+  <title>Mint Components Stencilbook</title>
+  <script type="module" src="build/mint-components.esm.js"></script>
+  <script nomodule src="build/mint-components.js"></script>
+</head>
 
-    <sqb-auth-template-switch verified>
-      <template slot="logged-out">
-        <sqm-widget-verification></sqm-widget-verification>
-      </template>
-      <template slot="logged-in">
-        <h1>VERIFIED</h1>
-      </template>
-    </sqb-auth-template-switch> -->
-  </body>
+<body>
+  <sqm-stencilbook></sqm-stencilbook>
+</body>
+
 </html>
